requirements.txt
pyjwt
flask-jwt-extended
Flask-Bcryp
Flask-PyMongo
exponent_server_sdk
flasgger
